Remove `MemoBytes` from `ShelleyTx`
