+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Tx.hs
@@ -31,7 +31,6 @@ import Cardano.Ledger.Shelley.Tx (
   mkBasicShelleyTx,
   shelleyMinFeeTx,
   sizeShelleyTxF,
-  wireSizeShelleyTxF,
   witsShelleyTxL,
  )
 import qualified Data.Set as Set (map)
@@ -55,9 +54,6 @@ instance EraTx AllegraEra where
 
   sizeTxF = sizeShelleyTxF
   {-# INLINE sizeTxF #-}
-
-  wireSizeTxF = wireSizeShelleyTxF
-  {-# INLINE wireSizeTxF #-}
 
   validateNativeScript = validateTimelock
   {-# INLINE validateNativeScript #-}
